check if README is generated

+++ nbprocess/cli.py
@@ -282,8 +282,9 @@ def nbprocess_quarto(
     os.system(f'cd {path} && quarto render --no-execute')
     os.system(f'cd {path} && quarto render {files[-1]} -o README.md -t gfm --no-execute')
     
-    (cfg_path/'README.md').unlink(missing_ok=True)
-    shutil.move(tmp_doc_path/'README.md', cfg_path) # README.md is temporarily in the nbs/docs folder
+    if (tmp_doc_path/'README.md').exists():
+        (cfg_path/'README.md').unlink(missing_ok=True)
+        shutil.move(tmp_doc_path/'README.md', cfg_path) # README.md is temporarily in the nbs/docs folder
     
     if tmp_doc_path.parent != cfg_path: # move docs folder to root of repo if it doesn't exist there
         shutil.rmtree(doc_path, ignore_errors=True)
